Revert "Testing new login button on deployed branch"
